+++ server/routes/index.js
@@ -1,0 +1,13 @@
+var express = require('express');
+var router = express.Router();
+
+/* GET home page. */
+router.get('/', function (req, res, next) {
+  // View-Model
+  res.render('index', {
+    title: 'Express',
+    author: 'Gonzalo Hazel'
+  });
+});
+
+module.exports = router;
