+++ server/views/index.hbs
@@ -1,2 +1,3 @@
 <h1>{{title}}</h1>
 <p>Welcome to {{title}}</p>
+<p> 👨‍💻 Autor: {{author}} </p>
